requirements.txt
# Core Dependencies
flet>=0.21.0
appwrite>=4.0.0
python-dotenv>=1.0.0

# Development Dependencies
pyinstaller>=5.0.0
pytest>=7.0.0
flake8>=6.0.0

# Web API Dependencies
flask>=2.3.0
flask-cors>=4.0.0

# Optional Dependencies for enhanced features
requests>=2.28.0
cryptography>=3.4.8
pillow>=9.0.0
plotly>=5.17.0
pandas>=2.1.0
